+++ internal/infra/push/datadestination_oracle.go
@@ -32,85 +32,66 @@ func (d OracleDialect) Placeholder(position int) string {
 	return fmt.Sprintf(":v%d", position)
 }
 
-// EnableConstraintsStatement generate statments to activate constraintes
-func (d OracleDialect) EnableConstraintsStatement(tableName string) string {
-	tableComposants := strings.Split(tableName, ".")
-	switch len(tableComposants) {
-	case 1:
-		return fmt.Sprintf(`BEGIN
-	FOR c IN
-	(SELECT c.owner, c.table_name, c.constraint_name
-	 FROM user_constraints c, user_tables t
-	 WHERE c.table_name = t.table_name
-	 AND c.owner = sys_context( 'userenv', 'current_schema' )
-	 AND c.table_name = '%s'
-	 AND c.status = 'DISABLED'
-	 ORDER BY c.constraint_type)
-	LOOP
-	  dbms_utility.exec_ddl_statement('alter table "' || c.owner || '"."' || c.table_name || '" enable constraint ' || c.constraint_name);
-	END LOOP;
-  END;
-  `, tableName)
-
-	case 2:
-		return fmt.Sprintf(`BEGIN
-		FOR c IN
-		(SELECT c.owner, c.table_name, c.constraint_name
-		 FROM user_constraints c, user_tables t
-		 WHERE c.table_name = t.table_name
-		 AND c.owner = '%s'
-		 AND c.table_name = '%s'
-		 AND c.status = 'DISABLED'
-		 ORDER BY c.constraint_type)
-		LOOP
-		  dbms_utility.exec_ddl_statement('alter table "' || c.owner || '"."' || c.table_name || '" enable constraint ' || c.constraint_name);
-		END LOOP;
-	  END;
-	  `, tableComposants[0], tableComposants[0])
-	default:
-		return ""
+// DisableConstraintsStatement generate statement to deactivate constraints
+func (d OracleDialect) DisableConstraintsStatement(tableName string) string {
+	schemaAndTable := strings.Split(tableName, ".")
+	sql := &strings.Builder{}
+	sql.WriteString(
+		`BEGIN
+		 FOR c IN(
+		 SELECT c.owner, c.table_name, c.constraint_name
+		 FROM user_constraints c
+		 CONNECT BY PRIOR c.constraint_name = c.r_constraint_name
+		 START WITH c.constraint_name IN (
+			SELECT c.constraint_name
+			FROM user_constraints c
+		 	WHERE c.status = 'ENABLED' AND c.table_name = '`)
+	if len(schemaAndTable) == 2 {
+		sql.WriteString(schemaAndTable[1])
+		sql.WriteString("' AND c.owner = '")
+		sql.WriteString(schemaAndTable[0])
+		sql.WriteString("'")
+	} else {
+		sql.WriteString(schemaAndTable[0])
+		sql.WriteString("' AND c.owner = sys_context( 'userenv', 'current_schema' )")
 	}
+	sql.WriteString(`)
+		LOOP
+			dbms_utility.exec_ddl_statement('alter table "' || c.owner || '"."' || c.table_name || '" disable constraint ' || c.constraint_name);
+		END LOOP;
+	END;`)
+	return sql.String()
 }
 
-// DisableConstraintsStatement generate statments to deactivate constraintes
-func (d OracleDialect) DisableConstraintsStatement(tableName string) string {
-	tableComposants := strings.Split(tableName, ".")
-	switch len(tableComposants) {
-	case 1:
-		return fmt.Sprintf(`BEGIN
-	FOR c IN
-		(SELECT c.owner, c.table_name, c.constraint_name
-		FROM user_constraints c, user_tables t
-		WHERE c.table_name = t.table_name
-		AND c.owner = sys_context( 'userenv', 'current_schema' )
-		AND c.table_name = '%s'
-		AND c.status = 'ENABLED'
-		AND NOT (t.iot_type IS NOT NULL AND c.constraint_type = 'P')
-		ORDER BY c.constraint_type DESC)
-	LOOP
-		dbms_utility.exec_ddl_statement('alter table "' || c.owner || '"."' || c.table_name || '" disable constraint ' || c.constraint_name);
-	END LOOP;
-  END;
-  `, tableName)
-	case 2:
-		return fmt.Sprintf(`BEGIN
-	FOR c IN
-		(SELECT c.owner, c.table_name, c.constraint_name
-		FROM user_constraints c, user_tables t
-		WHERE c.table_name = t.table_name
-		AND c.owner = '%s'
-		AND c.table_name = '%s'
-		AND c.status = 'ENABLED'
-		AND NOT (t.iot_type IS NOT NULL AND c.constraint_type = 'P')
-		ORDER BY c.constraint_type DESC)
-	LOOP
-		dbms_utility.exec_ddl_statement('alter table "' || c.owner || '"."' || c.table_name || '" disable constraint ' || c.constraint_name);
-	END LOOP;
-	END;
-	`, tableComposants[0], tableComposants[1])
-	default:
-		return ""
+// EnableConstraintsStatement generate statments to activate constraintes
+func (d OracleDialect) EnableConstraintsStatement(tableName string) string {
+	schemaAndTable := strings.Split(tableName, ".")
+	sql := &strings.Builder{}
+	sql.WriteString(
+		`BEGIN
+		 FOR c IN(
+		 SELECT c.owner, c.table_name, c.constraint_name
+		 FROM user_constraints c
+		 CONNECT BY PRIOR c.constraint_name = c.r_constraint_name
+		 START WITH c.constraint_name IN (
+			SELECT c.constraint_name
+			FROM user_constraints c
+		 	WHERE c.status = 'DISABLED' AND c.table_name = '`)
+	if len(schemaAndTable) == 2 {
+		sql.WriteString(schemaAndTable[1])
+		sql.WriteString("' AND c.owner = '")
+		sql.WriteString(schemaAndTable[0])
+		sql.WriteString("'")
+	} else {
+		sql.WriteString(schemaAndTable[0])
+		sql.WriteString("' AND c.owner = sys_context( 'userenv', 'current_schema' )")
 	}
+	sql.WriteString(`)
+		LOOP
+			dbms_utility.exec_ddl_statement('alter table "' || c.owner || '"."' || c.table_name || '" disable constraint ' || c.constraint_name);
+		END LOOP;
+	END;`)
+	return sql.String()
 }
 
 // TruncateStatement generate statement to truncat table content